% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closer_dwd_station.R
\name{closer_dwd_station}
\alias{closer_dwd_station}
\title{Information on closer DWD station}
\usage{
closer_dwd_station(
  lat,
  lon,
  start_date = "2023-01-01",
  end_date = "2023-01-31",
  stations_table,
  return_string = "id"
)
}
\arguments{
\item{lat}{(dbl) latitude in degree of point of interest}

\item{lon}{(dbl) longitude in degree of point of interest}

\item{start_date}{(chr) stations must have data from this date or earlier (e.g. 2022-06-30)}

\item{end_date}{(chr) stations must have data until this date or later (e.g. 2023-01-30)}

\item{stations_table}{(data.frame) table of stations info as provided by
functions \code{get_dwd_stations_info()} or \code{read_dwd_stations_info_file()}}

\item{return_string}{(chr) will return station id if set to "id",
name if set to "name", a string with all info separated by commas otherwise}
}
\value{
(chr) information about the station (see parameter return_string)
}
\description{
Identify the closest DWD station from a given point on earth and return
selected information about the station.
}
\examples{
# Read an example Stations Info file for Air Temperature
file.name.1 <- "TU_Stundenwerte_Beschreibung_Stationen.txt"
path.1 <- system.file("extdata", file.name.1, package = "zeppr")
my.stations <- read_dwd_stations_info_file(path.1)
closer_dwd_station(52.518611, 13.4083330, stations_table=my.stations)
closer_dwd_station(52.518611, 13.4083330, stations_table=my.stations, return_string = "name")
closer_dwd_station(52.518611, 13.4083330, stations_table=my.stations, return_string = "all")
}
