% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isip_weather_data.R
\name{create_isip_weather_data_input_table}
\alias{create_isip_weather_data_input_table}
\title{Creates an Input File for ISIP Weather Data Export Service}
\usage{
create_isip_weather_data_input_table(
  locations,
  latitudes,
  longitudes,
  start_date = NA,
  end_date = NA,
  daily = FALSE,
  output.excel.path = "isip.input.xlsx",
  overwrite = FALSE
)
}
\arguments{
\item{locations}{(chr) vector of location names}

\item{latitudes}{(dbl) vector of latitudes}

\item{longitudes}{(dbl) vector of longitudes}

\item{start_date}{(chr) starting date for weather data (format: "2021-12-31"; set to 2*365 days ago if NA)}

\item{end_date}{(chr) ending date for weather data (format: "2022-12-31"; set to 7 days ago if NA)}

\item{daily}{(boolean) returns daily data if TRUE, hourly data otherwise}

\item{output.excel.path}{(chr) path to the output Excel file}

\item{overwrite}{(boolean) overwrite file if true, no overwrite otherwise}
}
\description{
Creates an Excel file to be used as input for the ISIP Weather Data Export
Service.
}
\examples{
# Let us use an example dataset:
locs <- c("A", "B", "C")
lats <- c(50, 49, 48)
lons <- c(10, 9, 8)
# create_isip_weather_data_input_table(locs, lats, lons, "2022-01-01", "2022-01-10")
# create_isip_weather_data_input_table(locs, lats, lons, "2022-01-01", "2022-01-10", daily=T)
}
