% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_indices.R
\name{day_temp_index}
\alias{day_temp_index}
\title{Day Temperature Index}
\usage{
day_temp_index(t.min, t.max, index = "heating")
}
\arguments{
\item{t.min}{(numeric) vector of minimum day temperatures}

\item{t.max}{(numeric) vector of maximum day temperatures (same length as \code{t.min})}

\item{index}{(chr) type of temperature index to calculate in English or German.
One of the following in English:
"ice", "frost", "vegetation", "heating", "summer", "tropical", "desertic".
Or one of the corresponding in German:
"Eistag", "Frosttag", "Vegetationstag", "Heiztag", "Sommertag",
"Tropentag", or "Wuestentag".}
}
\value{
(boolean) vector of TRUE or FALSE values: TRUE if \code{t.min} or \code{t.max}
are compatible to the requested \code{index} definition, FALSE otherwise (same length as \code{t.min}).
}
\description{
Calculates weather temperature indices from daily weather data.
It is based on definitions provided by the German Weather Service.
The requested index can be in English or German.
}
\details{
The following indices can be calculated:
ice (\code{t.max}<0), frost (\code{t.min}<0), vegetation ((\code{t.max}+\code{t.min})/2)>=5),
heating ((\code{t.max}+\code{t.min})/2)<15), summer (\code{t.max}>=25),
tropical (\code{t.max}>=30) or desertic day (\code{t.max}>=35).

Sources in German:

https://de.wikipedia.org/wiki/Klimatologie

https://www.dwd.de/DE/service/lexikon/Functions/glossar.html?nn=103346&lv2=101334&lv3=101452
}
\examples{
day_temp_index(t.min=12, t.max=32, index="tropical")
day_temp_index(t.min=c(12, 15), t.max=c(15, 30), index="heating")
}
