% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_indices.R
\name{day_temperature_index}
\alias{day_temperature_index}
\title{Day Temperature Index}
\usage{
day_temperature_index(t.min, t.max, index = "cold")
}
\arguments{
\item{t.min}{(numeric) vector of minimum day temperatures}

\item{t.max}{(numeric) vector of maximum day temperatures (same length as \code{t.min})}

\item{index}{(chr) type of temperature index to calculate in English or German.
One of the following in English:
"ice", "frost", "vegetation", "heating", "summer", "tropical", "hot", "desertic".
Or one of the corresponding in German:
"Eistag", "Frosttag", "Vegetationstag", "Heiztag", "Sommertag", "Tropennacht",
"Heisser_Tag", or "Wuestentag".}
}
\value{
(boolean) vector of TRUE or FALSE values if  (same length as \code{t.min})
}
\description{
Calculates weather indices from daily weather data:
ice, frost, vegetation, heating, summer, tropical, hot or desertic day.
It is based on definitions provided by the German Weather Service (see sources).
}
\details{
Sources in German:
https://de.wikipedia.org/wiki/Klimatologie
https://www.dwd.de/DE/service/lexikon/Functions/glossar.html?nn=103346&lv2=101334&lv3=101452
}
\examples{
day_temperature_index(t.min=12, t.max=32, index="hot")
day_temperature_index(t.min=c(12, 15), t.max=c(15, 30), index="heating")
}
