% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closer_dwd_station.R
\name{dist_on_earth}
\alias{dist_on_earth}
\title{Distance in kilometers between 2 points on earth}
\usage{
dist_on_earth(lat1, lat2, lon1, lon2, output_unit = "km")
}
\arguments{
\item{lat1}{(dbl) latitude in degree of point 1}

\item{lat2}{(dbl) latitude in degree of point 2}

\item{lon1}{(dbl) longitude in degree of point 1}

\item{lon2}{(dbl) longitude in degree of point 2}

\item{output_unit}{(chr) set to "miles" to get the distance in miles, it is in kilometers otherwise}
}
\value{
(dbl) distance in kilometer
}
\description{
Calculates the distance in kilometers between 2 points on earth defined by
their degree coordinates (decimal) using the grand sphere method
(at the equator, 1 longitude degree = 111 km)
}
\examples{
dist_on_earth(48.856667, 52.518611, 2.3516670, 13.4083330) # Paris - Berlin
dist_on_earth(40.712778, 34.052222, 74.005833, 118.243611) # New York - LA
dist_on_earth(40.712778, 34.052222, 74.005833, 118.243611, output_unit="miles")
}
