% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_annotations.R
\name{download_and_unzip}
\alias{download_and_unzip}
\title{Downloads and optionally decompresses a ZIP file}
\usage{
download_and_unzip(target.url, out.dir, overwrite = FALSE, unzip = TRUE)
}
\arguments{
\item{target.url}{URL of the ZIP file}

\item{out.dir}{output directory (will be created if not exists)}

\item{overwrite}{download and overwrite existing file if TRUE, do nothing otherwise}

\item{unzip}{descompress the zip file if TRUE, do not decompress otherwise}
}
\description{
Downloads and optionally decompresses a ZIP file
}
\examples{
# # Download NUTS files
# my.nuts.url=
# "https://gisco-services.ec.europa.eu/distribution/v2/nuts/shp/NUTS_RG_01M_2021_4326.shp.zip"
# download_and_unzip(my.nuts.url, out.dir="tmp/nuts", overwrite=FALSE, unzip=TRUE)
# # Download GRID file
# my.grid.url="https://gisco-services.ec.europa.eu/grid/grid_100km_surf.gpkg"
# download_and_unzip(my.grid.url, out.dir="tmp/grids", overwrite=FALSE, unzip=FALSE)
}
