% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dwd_station_data.R
\name{get_dwd_station_data}
\alias{get_dwd_station_data}
\title{Get DWD Station Data from Internet}
\usage{
get_dwd_station_data(
  station_id = "00044",
  category = "air_temperature",
  timerange = "recent"
)
}
\arguments{
\item{station_id}{(char) DWD station ID as 5-character string e.g. "00044"}

\item{category}{(char) a metrics category from: air_temperature, cloud_type,
cloudiness, dew_point, extreme_wind, moisture, precipitation, pressure,
soil_temperature, solar, sun, visibility, weather_phenomena, wind, wind_synop}

\item{timerange}{(char) equal to "recent" (historical not supported)}
}
\value{
(data.frame) data frame of the stations data or NA if download
failed. The data frame has the following columns: station_id, timestamp,
and a variable number of columns depending on the file.
All columns are of type chr except timestamp of type POSIXct.
}
\description{
Download station data file from DWD recent data (not historical) for a
requested metrics category (e.g. air_temperature or dew_point).
List of categories
https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/hourly/
}
\examples{
head(get_dwd_station_data(station_id="00044", category="air_temperature"))
head(get_dwd_station_data(station_id="00044", category="dew_point"))
head(get_dwd_station_data(station_id="00183", category="solar"))
}
