% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dwd_stations_info.R
\name{get_dwd_stations_info}
\alias{get_dwd_stations_info}
\title{Get DWD Stations Info from Internet}
\usage{
get_dwd_stations_info(category = "air_temperature", timerange = "recent")
}
\arguments{
\item{category}{(\code{chr}) one of the following metrics category: air_temperature,
cloud_type, cloudiness, dew_point, extreme_wind, moisture, precipitation,
pressure, soil_temperature, solar, sun, visibility, weather_phenomena,
wind, wind_synop}

\item{timerange}{(\code{chr}) equal to either "recent" or "historical"}
}
\value{
(data.frame) data frame of the stations info with the following
columns: station_id (\code{chr}), start_date (\code{date}), end_date (\code{date}),
altitude (\code{dbl}), latitude (\code{dbl}), longitude (\code{dbl}), station_name (\code{chr}),
bundesland (\code{chr}), metrics (\code{chr})
}
\description{
Retrieve information on weather stations of the German Weather Service (DWD)
from its FTP site for a requested metrics category
(e.g. air_temperature or dew_point). The original data file is not saved
locally.
}
\examples{
get_dwd_stations_info("air_temperature")
get_dwd_stations_info("dew_point")
get_dwd_stations_info("solar")
get_dwd_stations_info("air_temperature", timerange="historical")
}
