% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emra_data.R
\name{get_emra_historical_weather}
\alias{get_emra_historical_weather}
\title{Retrieves interpolated weather data from the EMRA online service (emra.julius-kuehn.de).}
\usage{
get_emra_historical_weather(lat, lon, start_date = NA, end_date = NA)
}
\arguments{
\item{lat}{(dbl) latitude in degree of point of interest}

\item{lon}{(dbl) longitude in degree of point of interest}

\item{start_date}{(chr) Date formatted as YYYY-MM-DD (e.g. 2022-06-30).
The function will return historical data from this date if properly defined,
or from 1991-01-01 if \code{NA}}

\item{end_date}{(chr) Date formatted as YYYY-MM-DD (e.g. 2022-06-30).
The function will return historical data until this date if properly defined,
or until yesterday if \code{NA}}
}
\value{
(data.frame) a table of weather data defined by the following columns:
gridID, lat, lon, date, tmit (Tavg 2m), tmin (Tmin 2m), tmax (Tmax 2m),
emin (Tmin 5cm), tfmin (min Feuchttemperatur 2m), rrsum (sum precipitation mm),
rfmit (avg relative humidity \%), rrmax (max hourly precipitation mm),
rrgt10mm (n hours with precipitation > 10mm), radolansum (precipitation mm),
radolanmax (max hourly precipitation mm), radolangt10mm (n hours with precipitation > 10mm),
windmit (avg wind speed), rgmax (max hourly radiation), rgmit (avg hourly radiation),
sundur (sunlight duration h), bKultur01,
etpKultur01, etaKultur01, tschale, tapfel, tblatt
See full documentation at
https://emra.julius-kuehn.de/dokumente/upload/b916d_EMRA_Dokumentation__Webdienste.pdf
}
\description{
Historical daily data is available from 1991 to yesterday.
Predicted hourly data is provided for today and the next 7 days.
Geographic coordinates must be from projection PSG::31467 (https://spatialreference.org/).
The daily and hourly data is provided as raster data. A raster has an area of
1 x 1km2 and is defined within a grid over Germany.
}
\examples{
  lon <- 13.211756
  lat <- 52.292031
  before_yesterday <- format(Sys.Date()-2,"\%Y-\%m-\%d")
  # my.table <- get_emra_historical_weather(lat, lon, start_date=before_yesterday)
  # head(my.table[,1:8])
  # my.table <- get_emra_historical_weather(lat, lon, end_date="1991-01-02")
  # head(my.table[,1:8])
}
