% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isip_data.R
\name{get_isip_charts_token}
\alias{get_isip_charts_token}
\title{Gets Token from ISIP Charts API}
\usage{
get_isip_charts_token(
  user,
  passwd,
  url,
  path = NA,
  file_name = "ISIP.charts.token.txt",
  verbose = FALSE
)
}
\arguments{
\item{user}{(chr) API user name}

\item{passwd}{(chr) API password}

\item{url}{(chr) API URL}

\item{path}{(chr) directory path to store the token for reuse,
the working directory if \code{path} is \code{NA}.}

\item{file_name}{(chr) output file name (default: ISIP.charts.token.txt)}

\item{verbose}{(bool) prints messages if TRUE, nothing otherwise}
}
\value{
(chr) a 24h-valid token
}
\description{
Retrieves online a token to access the ISIP Charts API.
If already locally stored, reloads and returns an existing token.
Valid credentials are required to get a token!
}
\examples{
  \dontrun{
    # Required credentials and URL are stored in a local file
    secrets <- read.delim("secrets.tsv", sep="\t")
    user <- secrets[secrets$key=="charts_api_user",]$value
    passwd <- secrets[secrets$key=="charts_api_pass",]$value
    charts_api_token_url <- secrets[secrets$key=="charts_api_token_url",]$value
    token <- get_isip_charts_token(
      user=user, passwd=passwd, url=charts_api_token_url, verbose = TRUE)
  }
}
