% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jansen_data.R
\name{get_jansen_historical_weather}
\alias{get_jansen_historical_weather}
\title{Gets JANSEN Interpolated Daily Weather Data}
\usage{
get_jansen_historical_weather(
  lat,
  lon,
  start_date = NA,
  end_date = NA,
  url,
  retries = 3,
  sleep_time = 3
)
}
\arguments{
\item{lat}{(dbl) latitude in degree of point of interest}

\item{lon}{(dbl) longitude in degree of point of interest}

\item{start_date}{(chr) Date formatted as YYYY-MM-DD (e.g. 2022-06-30).
The function will return historical data from this date if properly defined,
or from 1991-01-01 if \code{NA} (1991-01-01 is the first available date)}

\item{end_date}{(chr) Date formatted as YYYY-MM-DD (e.g. 2022-06-30).
The function will return historical data until this date if properly defined,
or until yesterday if \code{NA} (the date of yesterday is the last available date)}

\item{url}{(chr) API URL}

\item{retries}{(int) number of times to retry if an error occurs during download}

\item{sleep_time}{(int) number of seconds between 2 retries}
}
\value{
(data.frame) a table of daily weather data defined by the following columns:
gridID (chr) (JKI grid ID),
lat (dbl) (latitude),
lon (dbl) (longitude),
date (date) (YYYY-MM-DD),
temperature (dbl) (average air temperature, 째C),
temperatureMin (dbl) (minimum air temperature, 째C),
temperatureMax (dbl) (maximum air temperature, 째C),
precipitation (dbl) (sum precipitation, mm),
vpgsl (dbl) (potential evaporation over grass, mm)
zsbnas (dbl) (hours of leaf wetness, h)
zsu70 (dbl) (hours above 70\% relative humidity, h),
zsu80 (dbl) (hours above 80\% relative humidity, h),
zsu90 (dbl) (hours above 90\% relative humidity, h),
windspeed (dbl) (daily average wind speed, m/s),
globalRadiation (dbl) (daily average global radiation, J/cm2),
tmu (dbl) (daily average relative humidity, \%),
sd (dbl) (sum sunlight duration, h),
tmtd (dbl) (daily average dew point, 째C),
radolangt10mm (dbl) (hours with precipitation from RADOLAN > 10mm, h),
radolanmax (dbl) (max precipitation from RADOLAN, mm),
radolansum (dbl)  (sum precipitation from RADOLAN, mm),
bf_kultur_01 (dbl) (sum soil humidity 0-60cm under winter wheat, \%),
etp_kultur_01 (dbl) (sum of potential evaporation under winter wheat, mm),
eta_kultur_01 (dbl) (sum of current (or real?) evaporation under winter wheat, mm),
rr24 (dbl) (24h precipitation sum, from yesterday to today 6:00 UTC, mm)
}
\description{
JANSEN is an online service (synops.julius-kuehn.de).
Daily data is available from 1990 to yesterday.
The daily and hourly data is provided as raster data. A raster has an area of
1 x 1km2 and is defined within a grid over Germany.
}
\examples{
  \dontrun{
    # Input parameters
    secrets <- read.delim("secrets.tsv", sep="\t")
    jansen_api_url <- secrets[secrets$key=="jansen_api_url",]$value
    before_yesterday <- format(Sys.Date()-2,"\%Y-\%m-\%d")
    # Requests
    get_jansen_historical_weather(52.292031, 13.211756, start_date=before_yesterday, url=jansen_api_url)
    get_jansen_historical_weather(52.292031, 13.211756, end_date="1990-01-03", url=jansen_api_url)
  }
}
