% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jansen_data.R
\name{get_jki_gridid}
\alias{get_jki_gridid}
\title{Derives JKI internal GridIDs from given geographic points}
\usage{
get_jki_gridid(x, y, crs = 4326)
}
\arguments{
\item{x}{vector of x coordinates of the input geographic points}

\item{y}{vector of y coordinates of the input geographic points}

\item{crs}{(int/chr) coordinate reference system of \code{x} and \code{y}: integer with the EPSG code,
or character with proj4stringcrs of the input coordinates (default: 4326 for WGS 84)}
}
\value{
(chr) JKI internal GridIDs
}
\description{
A JKI internal GridID for a given geographic point is a string composed
of the first 4 digits of the X and Y coordinates in the projection EPSG::31467.
It is used in various online services of JKI.
}
\examples{
  get_jki_gridid(x=13.211756, y=52.292031) # expecting: 37875803
}
