% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growing_degree_days.R
\name{growing_degree_days}
\alias{growing_degree_days}
\title{Growing Degree-Days}
\usage{
growing_degree_days(
  t.min,
  t.max,
  t.ceiling = 30,
  t.base = 5,
  use.floor = FALSE
)
}
\arguments{
\item{t.min}{(numeric) vector of minimum day temperatures}

\item{t.max}{(numeric) vector of maximum day temperatures (same length as \code{t.min})}

\item{t.ceiling}{(dbl) ceiling temperature value for \code{t.max} and \code{t.min}}

\item{t.base}{(dbl) base growth temperature}

\item{use.floor}{(boolean) \code{t.base} is used as a floor value
for \code{t.max} and \code{t.min} if \code{TRUE}, nothing more otherwise}
}
\value{
(numeric) vector of growing degree-days (same length as \code{t.min})
}
\description{
Calculates the growing degree-days independently for each pair of min
and max day temperature values.
Min and max Temperature values are provided as 2 vectors of same size.
A pair is defined as 2 values at same position in the vectors.
Although no time data is required here for the calculation,
each pair represents implicitly a time point: either 1 day or 1 hour.
Basic Formula: ((max temperature + min temperature)/2) - base temperature
}
\examples{
# As the following defines, the growing degree-days gdd = ((9+12)/2)-10 = 0.5
growing_degree_days(9, 12, t.ceiling=30, t.base=10, use.floor=FALSE)
# As the following defines, gdd = ((10+12)/2)-10 = 1
growing_degree_days(9, 12, t.ceiling=30, t.base=10, use.floor=TRUE)
# As the following defines, gdd = ((10+15)/2)-10 = 2.5
growing_degree_days(4, 20, t.ceiling=15, t.base=10, use.floor=TRUE)
# Processes pairs of min max temperatures defined in 2 vectors of same length
growing_degree_days(c(7, 8, 10), c(12, 14, 15),
  t.ceiling=30, t.base=10, use.floor=FALSE)
}
