% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_tool.R
\name{index_first_value_if}
\alias{index_first_value_if}
\title{Binary Indexes the first value in a vector complying to a condition.
The value is from a given numerical vector and the condition is tested by
comparing the vector values to a reference value using a common logical operator.}
\usage{
index_first_value_if(
  x,
  operator = "==",
  ref.value = 0,
  complete.to.end = FALSE
)
}
\arguments{
\item{x}{(numeric) vector of numeric values}

\item{operator}{(chr) type of comparison to perform among: "==", "<", "<=", ">", ">=", "is.na"}

\item{ref.value}{(numeric) the reference value}

\item{complete.to.end}{(bool) change the output by returning TRUE for the
first value complying to the condition and also for all values after the first,
FALSE otherwise}
}
\value{
(boolean) vector (same length as \code{x}): TRUE for the first value complying to the condition, FALSE otherwise
}
\description{
Binary Indexes the first value in a vector complying to a condition.
The value is from a given numerical vector and the condition is tested by
comparing the vector values to a reference value using a common logical operator.
}
\examples{
x <- c(1, NA, 7, 8, 4, 1, 2, NA, 4)
index_first_value_if(x=x, operator=">", ref.value=6, complete.to.end=FALSE)
index_first_value_if(x=x, operator=">", ref.value=6, complete.to.end=TRUE)
index_first_value_if(x=x, operator="is.na")
}
