% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_annotations.R
\name{map_nuts_name}
\alias{map_nuts_name}
\title{gets European administrative area names for given geographic points.
Source 2021 shape files to be downloaded from
https://gisco-services.ec.europa.eu/distribution/v2/nuts/shp/NUTS_RG_01M_2021_4326.shp.zip
and decompressed}
\usage{
map_nuts_name(lat, lon, level = 0, german.iso.codes = FALSE, nuts.path)
}
\arguments{
\item{lat}{(numeric) vector of latitudes of the geographic point}

\item{lon}{(numeric) vector of longitudes of the geographic point}

\item{level}{(numeric) NUTS annotation level equal to 0, 1, 2 or 3.
Depending on the country, the level fits approximately to the following entities:
0 for country, 1 for region/state, 2 for district/state, 3 for communalities/district}

\item{german.iso.codes}{(boolean) for NUTS level 1 in Germany, returns the 2-letter German ISO codes if TRUE, full federal state names otherwise}

\item{nuts.path}{(chr) local non-compressed NUTS shape file path (e.g. NUTS_RG_01M_2021_4326.shp)}
}
\description{
gets European administrative area names for given geographic points.
Source 2021 shape files to be downloaded from
https://gisco-services.ec.europa.eu/distribution/v2/nuts/shp/NUTS_RG_01M_2021_4326.shp.zip
and decompressed
}
\examples{
# # Download 2021 NUTS files in temporary folder
# my.nuts.url=
# "https://gisco-services.ec.europa.eu/distribution/v2/nuts/shp/NUTS_RG_01M_2021_4326.shp.zip"
# download_and_unzip(my.nuts.url, out.dir="tmp/nuts", overwrite=FALSE, unzip=TRUE)
# # Use local NUTS files
# my.nuts.path="tmp/nuts/NUTS_RG_01M_2021_4326.shp"
# map_nuts_name(lat=49.991, lon=8.256, level=0, nuts.path=my.nuts.path)
# map_nuts_name(lat=49.991, lon=8.256, level=1, german.iso.codes=FALSE, nuts.path=my.nuts.path)
# map_nuts_name(lat=49.991, lon=8.256, level=1, german.iso.codes=TRUE, nuts.path=my.nuts.path)
# map_nuts_name(lat=50.96, lon=10.78, level=1, german.iso.codes=TRUE, nuts.path=my.nuts.path)
# map_nuts_name(lat=48.42, lon=8.96, level=1, german.iso.codes=TRUE, nuts.path=my.nuts.path)
}
