% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growing_degree_days.R
\name{mutate_cumsum_gdd}
\alias{mutate_cumsum_gdd}
\title{Adds Growing Degree-Days Column to a Data Frame}
\usage{
mutate_cumsum_gdd(
  df,
  date,
  t.min,
  t.max,
  t.ceiling = 30,
  t.base = 5,
  use.floor = FALSE,
  hourly.data = FALSE,
  values_to = "cumsum_dd"
)
}
\arguments{
\item{df}{(data.frame) table with columns for date (or date including time for hourly data), min and max temperatures}

\item{date}{non-quoted column name containing dates (\code{date} column type)}

\item{t.min}{non-quoted column name containing minimum day temperatures (numeric column)}

\item{t.max}{non-quoted column name containing maximum day temperatures (numeric column)}

\item{t.ceiling}{(dbl) ceiling temperature value for \code{t.max} and \code{t.min}}

\item{t.base}{(dbl) base temperature}

\item{use.floor}{(boolean) if \code{TRUE}, \code{t.base} is a floor value for \code{t.max} and \code{t.min}}

\item{hourly.data}{(boolean) if \code{TRUE} considers dates as hourly data (divides the output by 24),
as daily data otherwise}

\item{values_to}{(chr) name of the new column to store the output}
}
\value{
(data.frame) table copying \code{df} but adding a numerical column for the cumulative
sum of growing degree-days
}
\description{
The input data frame must include columns for date, min and max
temperatures. Calculations are done by groups in grouped data frames.
Thus, the data.frame should be grouped by relevant variables such as
years and geographical locations otherwise it is assumed that there is only
1 location.
In each group of the data frame or in the full data frame if no groups are
defined, every time point from 1st January to last date of interest
must be represented by exactly one row (no missing time
point). \code{NA}s are not supported for the min and max temperatures.
Then, the function calculates the growing degree-days independently for
each row as defined in \code{\link[=growing_degree_days]{growing_degree_days()}}.
Finally it calculates the cumulative sum in a new column.
}
\examples{
# given a simple data.frame with date, min and max temperature
data <- data.frame(
  Date=as.Date(c("2022-01-01", "2022-01-02", "2022-01-02")),
  Tmin=c(4, 6, 11),
  Tmax=c(12, 14, 20))
# we add a column containing growing degree-days as follows:
mutate_cumsum_gdd(data, Date, Tmin, Tmax)
# With pipes from magrittr:
library(magrittr)
data \%>\% mutate_cumsum_gdd(Date, Tmin, Tmax)
}
