% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_indices.R
\name{mutate_day_temperature_index}
\alias{mutate_day_temperature_index}
\title{Adds Columns for Daily Temperature Indices to a Data Frame}
\usage{
mutate_day_temperature_index(
  df,
  t.min,
  t.max,
  german.output = FALSE,
  indices = c("ice", "frost", "vegetation", "heating", "summer", "tropical", "hot",
    "desertic")
)
}
\arguments{
\item{df}{(data.frame) table with columns for min and max day temperatures}

\item{t.min}{non-quoted column name containing minimum day temperatures (numeric column)}

\item{t.max}{non-quoted column name containing maximum day temperatures (numeric column)}

\item{german.output}{(boolean) returns column names in German if TRUE, in English otherwise}

\item{indices}{(chr vector) vector of index names in English to add to the
data frame. All available indices are added by Default. Available indices:
"ice", "frost", "vegetation", "heating", "summer", "tropical", "hot", "desertic".}
}
\value{
(data.frame) table copying \code{df} but adding a boolean column for
each requested index. TRUE if the day temperatures match the index definition,
FALSE otherwise.
}
\description{
The input data frame must contain daily weather data and include columns for
min and max temperatures. A new column will be added for each requested
temperature index. The new columns will contain boolean values: TRUE if the
day temperatures match the index definition, FALSE otherwise.
}
\details{
It is based on definitions provided by the German Weather Service. Sources:

https://de.wikipedia.org/wiki/Klimatologie

https://www.dwd.de/DE/service/lexikon/Functions/glossar.html?nn=103346&lv2=101334&lv3=101452
}
\examples{
# given a simple data.frame with date, min and max day temperatures
data <- data.frame(
  Date=as.Date(c("2022-01-01", "2022-01-02", "2022-01-02")),
  Tmin=c(4, 6, 11),
  Tmax=c(12, 14, 20))
# we add columns for all or selected temperature indices:
mutate_day_temperature_index(data, Tmin, Tmax)
mutate_day_temperature_index(data, Tmin, Tmax, indices=c("frost", "heating"))
mutate_day_temperature_index(data, Tmin, Tmax, german.output=TRUE)
# With pipes from magrittr:
library(magrittr)
data \%>\% mutate_day_temperature_index(Tmin, Tmax)
}
