% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isip_weather_data.R
\name{mutate_isip_daily_weather_with_cumsum_gdd}
\alias{mutate_isip_daily_weather_with_cumsum_gdd}
\title{mutates ISIP daily weather data table to add the cumulative sum of growing degree-days}
\usage{
mutate_isip_daily_weather_with_cumsum_gdd(
  isip.df,
  t.ceiling = 30,
  t.base = 5,
  use.floor = FALSE,
  values_to = "cumsum_gdd"
)
}
\arguments{
\item{isip.df}{(data.frame) table of weather data created with function \code{\link[=read_isip_hourly_to_daily_weather_data]{read_isip_hourly_to_daily_weather_data()}}}

\item{t.ceiling}{(dbl) ceiling temperature value for the min and max temperatures}

\item{t.base}{(dbl) base temperature}

\item{use.floor}{(boolean) if \code{TRUE}, \code{t.base} is a floor value for the min and max temperatures}

\item{values_to}{(chr) name of the new column to be added}
}
\value{
(data.frame) mutated table with an additional column for the cumulative sum of the growing degree-days
}
\description{
The ISIP data should be loaded by function \code{\link[=read_isip_hourly_to_daily_weather_data]{read_isip_hourly_to_daily_weather_data()}}.
For each location and each year in the data, daily time points (rows of the
table) are expected to be complete starting from 1st January to last time
point of interest (e.g. every day from 01 January to 30 June).
The data will be ordered by location and date, temporarily grouped by
location and year, and then the cumulative sum of growing degree-days at
each time point (each day) will be stored in a new column.
Growing degree-day formula: ((max + min temperature)/2) - base temperature
}
\examples{
# Open example file:
file <- "20221215_isip_hourly_weather_data_export.xlsx"
path <- system.file("extdata", file, package = "zeppr")
d <- read_isip_hourly_to_daily_weather_data(path)
# Apply function with default parameters:
mutate_isip_daily_weather_with_cumsum_gdd(d)
# Apply 2 times the function with magrittr pipes and tuned parameters:
library(magrittr)
d \%>\%
  mutate_isip_daily_weather_with_cumsum_gdd(
    t.ceiling=26, t.base=10, use.floor=TRUE, values_to="cs_gdd")
}
