% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isip_weather_data.R
\name{mutate_isip_hourly_weather_with_cumsum_gdd}
\alias{mutate_isip_hourly_weather_with_cumsum_gdd}
\title{mutates ISIP hourly weather data table to add the cumulative sum of growing degree-days}
\usage{
mutate_isip_hourly_weather_with_cumsum_gdd(
  isip.df,
  t.ceiling = 30,
  t.base = 5,
  use.floor = FALSE,
  values_to = "cumsum_gdd",
  max_per_day = FALSE
)
}
\arguments{
\item{isip.df}{(data.frame) table of weather data created with function \code{\link[=read_isip_hourly_weather_data]{read_isip_hourly_weather_data()}}}

\item{t.ceiling}{(dbl) ceiling temperature value}

\item{t.base}{(dbl) base temperature}

\item{use.floor}{(boolean) if \code{TRUE}, \code{t.base} is a floor temperature value}

\item{values_to}{(chr) name of the new column to be added}

\item{max_per_day}{(boolean) outputs the maximum per day if \code{TRUE}, reports for each hour otherwise}
}
\value{
(data.frame) mutated table with an additional column for the cumulative sum of the growing degree-days
}
\description{
The ISIP data should be loaded by function \code{\link[=read_isip_hourly_weather_data]{read_isip_hourly_weather_data()}}.
For each location and each year in the data, hourly time points (rows of the
table) are expected to be complete starting from 1st January to last time
point of interest (e.g. every hour from 01 January to 30 June).
The data will be ordered by location and time, temporarily grouped by
location and year, and then the cumulative sum of growing degree-days at
each time point (each hour) will be stored in a new column.
Growing degree-day formula: ((max + min temperature)/2) - base temperature
}
\examples{
# Open example file:
path <- system.file(
  "extdata",
  "20221215_isip_hourly_weather_data_export.xlsx",
  package = "zeppr")
d <- read_isip_hourly_weather_data(path)
# Apply function with default parameters:
mutate_isip_hourly_weather_with_cumsum_gdd(d)
# Apply 2 times the function with magrittr pipes and tuned parameters:
library(magrittr)
d \%>\%
  mutate_isip_hourly_weather_with_cumsum_gdd(
    t.ceiling=26, t.base=6, use.floor=TRUE, values_to="cs_gdd_1") \%>\%
  mutate_isip_hourly_weather_with_cumsum_gdd(
    t.ceiling=26, t.base=6, use.floor=TRUE, values_to="cs_gdd_2",
    max_per_day=TRUE) \%>\%
  print(n=48)
}
