% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isip_weather_data.R
\name{mutate_isip_weather_with_cumsum_gdd}
\alias{mutate_isip_weather_with_cumsum_gdd}
\title{Mutates ISIP Weather Table to Add the Cumulative Sum of Growing Degree-Days}
\usage{
mutate_isip_weather_with_cumsum_gdd(
  isip.df,
  t.ceiling = 30,
  t.base = 5,
  use.floor = FALSE,
  values_to = "cumsum_gdd",
  daily.data = FALSE,
  max_per_day = FALSE
)
}
\arguments{
\item{isip.df}{(data.frame) table of weather data created with
function \code{\link[=read_isip_hourly_weather_data]{read_isip_hourly_weather_data()}}}

\item{t.ceiling}{(dbl) ceiling temperature value}

\item{t.base}{(dbl) base temperature (base growth temperature)}

\item{use.floor}{(boolean) \code{t.base} is also a floor temperature value
if \code{TRUE}, nothing more otherwise}

\item{values_to}{(chr) name of the new column to be added}

\item{daily.data}{(boolean) process data as daily data if \code{TRUE}, as hourly
data otherwise. Processing hourly data as daily data will likely result in
only NA values in the new column. The other way around will return values
but they will not be correct.}

\item{max_per_day}{(boolean) outputs the maximum per day if \code{TRUE},
reports for each hour otherwise}
}
\value{
(data.frame) mutated table with an additional column for the
cumulative sum of the growing degree-days
}
\description{
The ISIP data should be loaded by function \code{\link[=read_isip_hourly_weather_data]{read_isip_hourly_weather_data()}}.
For each location and each year in the data, hourly time points (rows of the
table) are expected to be complete starting from 1st January to last time
point of interest (e.g. every hour from 01 January to 30 June).
The data will be ordered by location and time, temporarily grouped by
location and year, and then the cumulative sum of growing degree-days at
each time point (each hour) will be stored in a new column.
Growing degree-day formula: ((max + min temperature)/2) - base temperature
}
\examples{
# ---------------------------------------------------------------------------
# Choose an example file with hourly weather data:
# ---------------------------------------------------------------------------
file <- "20221215_isip_hourly_weather_data_export.xlsx"
path <- system.file("extdata", file, package = "zeppr")
# ---------------------------------------------------------------------------
# Reads hourly data and derive results by hours with or without day summary:
# ---------------------------------------------------------------------------
# reads file and keep 10 rows (10 hours) only
hourly.table <- read_isip_hourly_weather_data(path)[1:10,]
# adds cumulative growing degree-days per row divided by 24
mutate_isip_weather_with_cumsum_gdd(hourly.table)
# keep maximum value per day
mutate_isip_weather_with_cumsum_gdd(hourly.table, max_per_day=TRUE)
# ---------------------------------------------------------------------------
# Reads as daily data and derive results by days with or without floor value:
# ---------------------------------------------------------------------------
# reads file and keep 10 rows (10 days) only
daily.table <- read_isip_hourly_weather_data(path, returns.daily.data=TRUE)[1:10,]
# adds cumulative growing degree-days per row
mutate_isip_weather_with_cumsum_gdd(daily.table, daily.data=TRUE)
# use floor value for min and max temperatures
mutate_isip_weather_with_cumsum_gdd(daily.table, daily.data=TRUE, use.floor=TRUE)
# ---------------------------------------------------------------------------
# Apply 2 times the function with magrittr pipes and tuned parameters:
# ---------------------------------------------------------------------------
library(magrittr)
hourly.table <- read_isip_hourly_weather_data(path)[1:48,]
hourly.table \%>\%
  mutate_isip_weather_with_cumsum_gdd(
    t.ceiling=26, t.base=6, use.floor=TRUE, values_to="cs_gdd_1") \%>\%
  mutate_isip_weather_with_cumsum_gdd(
    t.ceiling=26, t.base=6, use.floor=TRUE, values_to="cs_gdd_2",
    max_per_day=TRUE) \%>\%
  print(n=48)
}
