% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_cumsum.R
\name{normalized_cumsum}
\alias{normalized_cumsum}
\title{Normalized Cumulative Sum}
\usage{
normalized_cumsum(values, normalize = TRUE, na.replace = NA)
}
\arguments{
\item{values}{(numeric) vector of numerical values}

\item{normalize}{(boolean) returns normalized cumulative sum if \code{TRUE}, simple cumulative sum otherwise}

\item{na.replace}{(dbl) numeric value used to replace \code{NA} values (default=\code{NA} to not replace)}
}
\value{
(numeric) vector of the cumulative sum normalized by default from 0 to 1
}
\description{
First The cumulative sum includes all values starting from the 1st element.
For example, cumulative sum of element i is equal to
sum(element 1, element 2, ... element i). Then, each value is normalized as a
percentage by dividing the value by the sum of all elements.
}
\examples{
# Simple case
normalized_cumsum(c(1,2,3,4,5))
# not-normalized cumulative sum
normalized_cumsum(c(1,2,3,4, 5), normalize=FALSE)
# NAs are not processed by default
normalized_cumsum(c(1,2,3,4, NA, 5))
# replace NAs by 0 before summing
normalized_cumsum(c(1,2,3,4, NA, 5), na.replace=0)
}
