% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dwd_stations_info.R
\name{read_dwd_stations_info_file}
\alias{read_dwd_stations_info_file}
\title{Read DWD Stations Info File}
\usage{
read_dwd_stations_info_file(path)
}
\arguments{
\item{path}{(\code{chr}) local file path to a DWD Stations Info file such as
"TU_Stundenwerte_Beschreibung_Stationen.txt"}
}
\value{
(\code{data.frame}) data frame of the stations info with the following
columns: station_id (\code{chr}), start_date (\code{date}), end_date (\code{date}),
altitude (\code{dbl}), latitude (\code{dbl}), longitude (\code{dbl}), station_name (\code{chr}),
bundesland (\code{chr})
}
\description{
Reads a local file describing weather stations of the German Weather
Service (DWD). The latest daily update of this file is available to download
from a subfolder of the following URL:
https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/hourly/.
For example, the stations info file related to air temperature will be found
from the URL above at air_temperature/recent/TU_Stundenwerte_Beschreibung_Stationen.txt
}
\examples{
# Read an example Stations Info file for Air Temperature
file.name.1 <- "TU_Stundenwerte_Beschreibung_Stationen.txt"
path.1 <- system.file("extdata", file.name.1, package = "zeppr")
head(read_dwd_stations_info_file(path.1))
# Read an example Stations Info file for Dew Point
file.name.2 <- "TD_Stundenwerte_Beschreibung_Stationen.txt"
path.2 <- system.file("extdata", file.name.2, package = "zeppr")
head(read_dwd_stations_info_file(path.2))
}
