% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isip_weather_data.R
\name{read_isip_hourly_to_daily_weather_data}
\alias{read_isip_hourly_to_daily_weather_data}
\title{Reads ISIP weather export data file (from hourly to daily data)}
\usage{
read_isip_hourly_to_daily_weather_data(excel.path)
}
\arguments{
\item{excel.path}{(chr) path to ISIP hourly weather export data file (Excel
file). ISIP's Excel Data format as of Dec 2022.}
}
\value{
(data.frame) table with columns: location (chr), date (dttm object),
Tmin, Tavg, Tmax, humidity, precipitation, radiation, wind_speed, n.hours
}
\description{
Reads ISIP hourly weather export data file (one row per hour), averages
values per day (except for min and max temperatures), and returns a data
frame with following columns of daily data (one row per day):
\code{date}, \code{location}, \code{Tmin} (minimum temperature),
\code{Tavg} (average temperature), \code{Tmax} (maximum temperature),
\code{humidity}, \code{precipitation}, \code{radiation}, \code{wind_speed},
\code{n.hours} (number of hours with data).
All sheets from the Excel file are processed and gathered in the same table.
Each Excel sheet contains the data for a particular geographical location
and the sheet name provides the location id after trimming a suffix
starting from the last underscore character.
}
\examples{
path <- system.file("extdata",
  "20221215_isip_hourly_weather_data_export.xlsx", package = "zeppr")
read_isip_hourly_to_daily_weather_data(path)
}
