% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isip_weather_data.R
\name{read_isip_weather_data}
\alias{read_isip_weather_data}
\title{Reads ISIP Weather Data File}
\usage{
read_isip_weather_data(
  excel.path,
  read.daily.data = FALSE,
  hourly.to.daily = FALSE,
  drop.irrelvant.cols = TRUE,
  col.names = c("ID", "date", "Tavg", "humidity", "precipitation", "radiation",
    "irrelevant", "wind_speed")
)
}
\arguments{
\item{excel.path}{(chr) path to ISIP weather export data file (Excel
file). ISIP's Excel Data format as of Dec 2022.}

\item{read.daily.data}{(boolean) input file is expected to contain daily data
if TRUE, hourly data otherwise}

\item{hourly.to.daily}{(boolean) returns daily data if \code{TRUE},
hourly data otherwise}

\item{drop.irrelvant.cols}{(boolean) remove irrelevant columns if \code{TRUE},
keep all columns otherwise (columns whose name start with 'irrelevant')}

\item{col.names}{(chr) vector of column names for the output table
(must match file content)}
}
\value{
(data.frame) table with numerical columns except for location (\code{chr})
and date (\code{Date} class for daily output or \code{POSIXct} class for hourly
output). Columns: \code{location}, \code{date}, \code{Tmin} (min daily temperature),
\code{Tavg} (average temperature), \code{Tmax} (max daily temperature), \code{humidity},
\code{precipitation}, \code{radiation}, \code{wind_speed} and
\code{n_hours} (scope of the data as number of hours, e.g. used to calculate averages).
}
\description{
Reads an ISIP hourly or daily weather export data file (Excel file;
one row per hour or per day, respectively)
and returns a data frame corresponding data. The function can convert hourly
to daily data. In this case, the daily data is derived from the hourly data
by averaging or summing up the values per day (sum for precipitation and
radiation, average otherwise). Minimum and maximum daily temperatures are
derived from this calculation and are not available directly from the file.
All sheets from the Excel file are processed and gathered in the same table.
Within an ISIP Excel file, each sheet represents a particular geographical
location and contains weather data for the same date (or time) range.
Sheet names provide the location ids after trimming a suffix
starting from the last underscore character.
}
\examples{
# Let us use an example hourly dataset:
file.name <- "20221215_isip_hourly_weather_data_export.xlsx"
path <- system.file("extdata", file.name, package = "zeppr")
# Read hourly data:
hourly.table <- read_isip_weather_data(path)
head(hourly.table)
# Read hourly data but summarise it as daily data:
daily.table <- read_isip_weather_data(path, hourly.to.daily=TRUE)
head(daily.table)
# Let us use an example daily dataset:
file.name.2 <- "20230330_isip_daily_weather_data_export.xlsx"
path.2 <- system.file("extdata", file.name.2, package = "zeppr")
# Read daily data:
daily.table.2 <- read_isip_weather_data(path.2, read.daily.data=TRUE)
head(daily.table.2)
}
