% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dwd_stations_info.R
\name{dwd_category_code}
\alias{dwd_category_code}
\title{Get DWD Category Code}
\usage{
dwd_category_code(category, lower_case = FALSE)
}
\arguments{
\item{category}{(\code{chr}) one of the following category: air_temperature,
cloud_type, cloudiness, dew_point, extreme_wind, moisture, precipitation,
pressure, soil_temperature, solar, sun, visibility, weather_phenomena,
wind, wind_synop}

\item{lower_case}{(\code{boolean}) output lowercase code if \code{TRUE}, upper case otherwise}
}
\value{
(\code{chr}) code corresponding to the supported categories, \code{NA} otherwise
}
\description{
Returns 1-2 letter code for a given metrics category defined on the German
Weather service (DWD) FTP site (e.g. TU for air_temperature, and TD
for dew_point).
}
\examples{
# zeppr:::dwd_category_code("air_temperature")
# zeppr:::dwd_category_code("air_temperature", lower_case=TRUE)
}
