% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dwd_stations_info.R
\name{dwd_category_code}
\alias{dwd_category_code}
\title{Get DWD Category Code}
\usage{
dwd_category_code(category)
}
\arguments{
\item{category}{(chr) choices: air_temperature, cloud_type, cloudiness,
dew_point, extreme_wind, moisture, precipitation, pressure, soil_temperature,
solar, sun, visibility, weather_phenomena, wind, wind_synop}
}
\value{
(chr) code corresponding to the supported categories, NA otherwise
}
\description{
Returns 1-2 letter code for a metrics category (e.g. TU for air_temperature,
and TD for dew_point).
}
\examples{
dwd_category_code("air_temperature")
dwd_category_code("dew_point")
}
