% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dwd_station_data.R
\name{get_dwd_station_data}
\alias{get_dwd_station_data}
\title{Get DWD Station Data from Internet}
\usage{
get_dwd_station_data(
  station_id = "00044",
  category = "air_temperature",
  timerange = "recent"
)
}
\arguments{
\item{station_id}{(\code{chr}) DWD station ID as 5-character string e.g. "00044"}

\item{category}{(\code{chr}) a metrics category from: air_temperature, cloud_type,
cloudiness, dew_point, extreme_wind, moisture, precipitation, pressure,
soil_temperature, solar, sun, visibility, weather_phenomena, wind, wind_synop}

\item{timerange}{(\code{chr}) either "recent" or "historical"}
}
\value{
(\code{data.frame}) data frame of the stations data or an empty data frame
if download failed. The data frame has the following columns:
station_id, timestamp, and a variable number of columns depending on the file.
All columns are of type chr except timestamp of type \code{POSIXct}.
}
\description{
Retrieve data of a weather station of the German Weather Service (DWD)
from its FTP site for a requested metrics category
(e.g. air_temperature or dew_point). The original data file is not saved
locally.
}
\examples{
head(get_dwd_station_data(station_id="04719", category="air_temperature", timerange="historical"))
head(get_dwd_station_data(station_id="00044", category="air_temperature"))
head(get_dwd_station_data(station_id="00183", category="solar"))
head(get_dwd_station_data(station_id="00044", category="dew_point", timerange="historical"))
}
