% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dwd_stations_info.R
\name{get_dwd_stations_info}
\alias{get_dwd_stations_info}
\title{Get DWD Stations Info from Internet}
\usage{
get_dwd_stations_info(category = "air_temperature", timerange = "recent")
}
\arguments{
\item{category}{(char) a metrics category from: air_temperature, cloud_type,
cloudiness, dew_point, extreme_wind, moisture, precipitation, pressure,
soil_temperature, solar, sun, visibility, weather_phenomena, wind, wind_synop}

\item{timerange}{(char) equal to "recent" (historical not supported)}
}
\value{
(data.frame) data frame of the stations info with the following
columns: station_id (chr), start_date (date), end_date (date), altitude (dbl),
latitude (dbl), longitude (dbl), station_name (chr), bundesland (chr),
metrics (chr)
}
\description{
Download stations info file from DWD recent data (not historical) for a
requested metrics category (e.g. air_temperature or dew_point).
List of categories
https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/hourly/
}
\examples{
get_dwd_stations_info("air_temperature")
get_dwd_stations_info("dew_point")
}
