% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emra_data.R
\name{get_emra_predicted_weather}
\alias{get_emra_predicted_weather}
\title{Gets EMRA Online Predicted Weather Data}
\usage{
get_emra_predicted_weather(lat, lon)
}
\arguments{
\item{lat}{(dbl) latitude in degree of point of interest}

\item{lon}{(dbl) longitude in degree of point of interest}
}
\value{
(data.frame) a table of weather data defined by the following columns:
date, doy (day of year), hourDB (hour), gridID, lat, lon,
tl (Tavg 2m), tf (Tmin 2m), rr (precipitation mm), vv (max wind speed m/s),
rg (max radiation J/cm),
bKultur01, etpKultur01, etaKultur01, tschale, tapfel, tblatt
See full documentation at
https://emra.julius-kuehn.de/dokumente/upload/b916d_EMRA_Dokumentation__Webdienste.pdf
}
\description{
Retrieves interpolated weather data from the EMRA online service (emra.julius-kuehn.de).
}
\details{
Historical daily data is available from 1991 to yesterday.
Predicted hourly data is provided for today and the next 7 days.
Geographic coordinates must be from projection PSG::31467 (https://spatialreference.org/).
The daily and hourly data is provided as raster data. A raster has an area of
1 x 1km2 and is defined within a grid over Germany.
}
\examples{
  lon <- 13.211756
  lat <- 52.292031
  # my.table <- get_emra_predicted_weather(lat, lon)
  # head(my.table[,1:8])
}
