% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isip_weather_data.R
\name{mutate_isip_weather_with_cumsum_gdd}
\alias{mutate_isip_weather_with_cumsum_gdd}
\title{Adds the Cumulative Sum of Growing Degree-Days to an ISIP Weather Data Table}
\usage{
mutate_isip_weather_with_cumsum_gdd(
  isip.df,
  t.ceiling = 30,
  t.base = 5,
  use.floor = FALSE,
  daily.data = FALSE,
  max.per.day = FALSE,
  values.to = "cumsum_gdd"
)
}
\arguments{
\item{isip.df}{(data.frame) table of weather data created with
function \code{\link[=read_isip_hourly_weather_data]{read_isip_hourly_weather_data()}}}

\item{t.ceiling}{(dbl) ceiling temperature value}

\item{t.base}{(dbl) base growth temperature}

\item{use.floor}{(boolean) \code{t.base} is also a floor temperature value
if \code{TRUE}, nothing more otherwise}

\item{daily.data}{(boolean) process data as daily data if \code{TRUE}, as hourly
data otherwise. Processing hourly data as daily data will likely result in
only NA values in the new column. The other way around will return values
but they will not be correct.}

\item{max.per.day}{(boolean) outputs the maximum per day if \code{TRUE},
reports for each hour otherwise (relevant only for hourly data)}

\item{values.to}{(chr) name of the new column to be added}
}
\value{
(data.frame) mutated table with an additional column for the
cumulative sum of the growing degree-days
}
\description{
The ISIP data should be loaded by function \code{\link[=read_isip_hourly_weather_data]{read_isip_hourly_weather_data()}}.
For each location and each year in the data, time points (rows of the
table) are expected to be complete starting from 1st January to last time
point of interest (e.g. every hour or every day from 01 January to 30 June).
The data will be ordered by location and time, temporarily grouped by
location and year, and then the cumulative sum of growing degree-days at
each time point (each day or each hour) will be stored in a new column.
Make sure that the \code{daily.data} parameter is appropriately set to fit
your data.
Growing degree-days are calculated as defined in function \code{\link[=growing_degree_days]{growing_degree_days()}}.
}
\examples{
# ---------------------------------------------------------------------------
# Choose an example file with hourly weather data:
# ---------------------------------------------------------------------------
file <- "20221215_isip_hourly_weather_data_export.xlsx"
path <- system.file("extdata", file, package = "zeppr")
# ---------------------------------------------------------------------------
# Reads hourly data and derive results by hours with or without day summary:
# ---------------------------------------------------------------------------
# reads file and keeps only 5 columns and 6 rows (6 hours)
hourly.table <- read_isip_hourly_weather_data(path)[1:6,1:5]
# adds cumulative growing degree-days per row divided by 24
mutate_isip_weather_with_cumsum_gdd(hourly.table)
# keep maximum value per day
mutate_isip_weather_with_cumsum_gdd(hourly.table, max.per.day=TRUE)
# ---------------------------------------------------------------------------
# Reads as daily data and derive results by days with or without floor value:
# ---------------------------------------------------------------------------
# reads file and keeps only 5 columns and 6 rows (6 days)
daily.table <- read_isip_hourly_weather_data(path, returns.daily.data=TRUE)[1:6,1:5]
# adds cumulative growing degree-days per row
mutate_isip_weather_with_cumsum_gdd(daily.table, daily.data=TRUE)
# use floor value for min and max temperatures
mutate_isip_weather_with_cumsum_gdd(daily.table, daily.data=TRUE, use.floor=TRUE)
# ---------------------------------------------------------------------------
# Apply 2 times the function with magrittr pipes and tuned parameters:
# ---------------------------------------------------------------------------
library(magrittr)
hourly.table <- read_isip_hourly_weather_data(path)[1:48,1:5]
hourly.table \%>\%
  mutate_isip_weather_with_cumsum_gdd(
    t.ceiling=26, t.base=6, use.floor=TRUE, values.to="cs_gdd_1") \%>\%
  mutate_isip_weather_with_cumsum_gdd(
    t.ceiling=26, t.base=6, use.floor=TRUE, values.to="cs_gdd_2",
    max.per.day=TRUE) \%>\%
  print(n=48)
}
