% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_trap_data.R
\name{normalize_trap_data}
\alias{normalize_trap_data}
\title{Normalizes Trap Data to Exposition Time}
\usage{
normalize_trap_data(
  df,
  date,
  values,
  default.duration = 7,
  use.mean.days = TRUE,
  values.to = "norm_val",
  ndays.to = NA,
  max.duration = NA
)
}
\arguments{
\item{df}{(data.frame) table of trap data containing at least a column for the
date and a column for the values}

\item{date}{non-quoted column name containing dates (\code{Date} class)}

\item{values}{non-quoted column name containing values (\code{dbl} class)}

\item{default.duration}{(integer) number of days defining the first interval
time duration}

\item{use.mean.days}{(boolean) the number of days defining the first interval
time duration is set to the mean duration of all other exposition times
if \code{TRUE}, to \code{default.duration} otherwise}

\item{values.to}{(chr) name of the new column to be added for the normalized values}

\item{ndays.to}{(chr) name of the new column to be added for the interval
time duration values in days, no column added if \code{NA}}

\item{max.duration}{(integer) maximal number of days between intervals if positive integer, no maximum otherwise}
}
\value{
(data.frame) mutated table with an additional column for the normalized
values, optionally with an other additional column for the interval
durations in days. Normalizing a NA value will result in a NA value.
Duplicated data points (multiple rows with same date within a group)
will produce Inf values.
}
\description{
Trap data are here numerical values recorded at different time points (dates).
Interval length between time points, the exposition time duration (in days),
may be irregular (e.g. every Mondays and Saturdays).
Traps are initialized at the beginning of each interval (value=0).
The end of each interval is a time point where values are recorded (e.g. insect counts).
This function normalizes the values as averages per day during the exposition time.
}
\details{
Important: the function can normalize values for multiple traps ONLY IF the
input data frame is grouped accordingly (using function \code{dplyr::group_by()}).
Otherwise (no groups), the function expects the data for a single trap.
In any case, the function will likely bug with NA values,
or with duplicate data points (multiple values associated
with the same time point). It is not necessary to order the data by date.
}
\examples{
# given a simple data.frame with date and counts
data <- data.frame(
  date=as.Date(c("2022-01-01", "2022-01-05", "2022-01-06", "2022-01-11")),
  counts=c(12, 14, 20, 12))
normalize_trap_data(data, date, counts)
# Using magrittr's pipes and keeping derived interval durations
library(magrittr)
data \%>\% normalize_trap_data(date, counts, ndays.to="n_days")
}
