% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dwd_stations_info.R
\name{read_dwd_stations_info_file}
\alias{read_dwd_stations_info_file}
\title{Read DWD Stations Info File}
\usage{
read_dwd_stations_info_file(path)
}
\arguments{
\item{path}{(chr) file path to a DWD Stations Info file such as
"TU_Stundenwerte_Beschreibung_Stationen.txt"}
}
\value{
(data.frame) data frame of the stations info with the following
columns: station_id, start_date, end_date, altitude, latitude,
longitude, station_name, bundesland
}
\description{
Reads a local file describing DWD stations downloaded from a subfolder of
the following URL
https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/hourly/
}
\examples{
# Read an example Stations Info file for Air Temperature
file.name.1 <- "TU_Stundenwerte_Beschreibung_Stationen.txt"
path.1 <- system.file("extdata", file.name.1, package = "zeppr")
head(read_dwd_stations_info_file(path.1))
# Read an example Stations Info file for Dew Point
file.name.2 <- "TD_Stundenwerte_Beschreibung_Stationen.txt"
path.2 <- system.file("extdata", file.name.2, package = "zeppr")
head(read_dwd_stations_info_file(path.2))
}
