% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isip_weather_data.R
\name{read_isip_hourly_weather_data}
\alias{read_isip_hourly_weather_data}
\title{Reads ISIP hourly weather export data file}
\usage{
read_isip_hourly_weather_data(excel.path)
}
\arguments{
\item{excel.path}{(chr) path to ISIP hourly weather export data file (Excel
file). ISIP's Excel Data format as of Dec 2022.}
}
\value{
(data.frame) with columns: location (chr),
date.time (dttm object), temperature, humidity, precipitation, radiation,
wind_speed
}
\description{
Reads ISIP hourly weather export data file (one row per hour) and returns
a data frame with following columns: \code{date.time}, \code{location},
\code{temperature}, \code{humidity}, \code{precipitation}, \code{radiation}, \code{wind_speed}.
All sheets from the Excel file are processed and gathered in the same table.
Each Excel sheet contains the data for a particular geographical location
and the sheet name provides the location id after trimming a suffix
starting from the last underscore character.
}
\examples{
path <- system.file("extdata",
  "20221215_isip_hourly_weather_data_export.xlsx", package = "zeppr")
read_isip_hourly_weather_data(path)
}
